requirements.txt
Django==2.2.5
django-taggit==1.1.0
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
